[Issue-51] Update AccountItem and AccountCard
