+++ components/theme/interface/components.ts
@@ -61,6 +61,7 @@ import type { ComponentToken as TypographyComponentToken } from '../../typograph
 import type { ComponentToken as UploadComponentToken } from '../../upload/style';
 import type { ComponentToken as TourComponentToken } from '../../tour/style';
 import type { ComponentToken as QRCodeComponentToken } from '../../qrcode/style';
+import type { ComponentToken as Web3BlockComponentToken } from '../../web3block/style';
 import type { ComponentToken as AppComponentToken } from '../../app/style';
 import type { ComponentToken as WaveToken } from '../../_util/wave/style';
 
@@ -141,6 +142,7 @@ export interface ComponentTokenMap {
   Progress?: ProgressComponentToken;
   Tour?: TourComponentToken;
   QRCode?: QRCodeComponentToken;
+  Web3Block?: Web3BlockComponentToken;
   App?: AppComponentToken;
 
   /** @private Internal TS definition. Do not use. */
