+++ components/web3block/account-item/stories/account-item.stories.tsx
@@ -1,41 +1,53 @@
 import React from 'react';
 import type { ComponentStory, ComponentMeta } from '@storybook/react';
-import { QrCode } from 'phosphor-react';
-import AccountCard from '..';
+import { Copy, QrCode } from 'phosphor-react';
+import AccountItem from '..';
 import Icon from '../../../icon';
-
-const { AccountCardSelection } = AccountCard;
+import Button from '../../../button';
+import Logo from '../../../logo';
 
 // More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 export default {
-  title: 'Core/AccountCardSelection',
-  component: AccountCardSelection,
+  title: 'SubWallet Component/AccountItem',
+  component: AccountItem,
   // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
   argTypes: {
     address: { control: 'text', defaultValue: '5HbcGs2QXVAc6Q6eoTzLYNAJWpN17AkCFRLnWDaHCiGYXvNc' },
-    accountName: { control: 'text', defaultValue: 'Account 5' },
+    size: { control: 'number', defaultValue: 40 },
     theme: { control: 'radio', options: ['polkadot', 'ethereum'], defaultValue: 'polkadot' },
     prefix: { control: 'number', defaultValue: 42 },
     preLength: { control: 'number', defaultValue: 9 },
     subLength: { control: 'number', defaultValue: 9 },
     isSelected: { control: 'boolean', defaultValue: false },
-    showCopy: { control: 'boolean', defaultValue: false },
-    showOption: { control: 'boolean', defaultValue: false },
+    isShowSubIcon: { control: 'boolean', defaultValue: false },
   },
-} as ComponentMeta<typeof AccountCardSelection>;
+} as ComponentMeta<typeof AccountItem>;
 
 // More on component templates: https://storybook.js.org/docs/react/writing-stories/introduction#using-args
-const Template: ComponentStory<typeof AccountCardSelection> = (args) => (
-  <AccountCardSelection {...args} />
-);
+const Template: ComponentStory<typeof AccountItem> = (args) => <AccountItem {...args} />;
 
-export const Default = Template.bind({});
+export const Selection = Template.bind({});
 // More on args: https://storybook.js.org/docs/react/writing-stories/args
-Default.args = {};
+Selection.args = {
+  subIcon: <Icon type="phosphor" phosphorIcon={QrCode} iconColor="#FFF" size="xs" weight="bold" />,
+  rightComponent: (
+    <Button
+      type="ghost"
+      size="xs"
+      icon={
+        <Icon type="phosphor" phosphorIcon={Copy} iconColor="rgba(255, 255, 255, 0.45)" size="sm" />
+      }
+    />
+  ),
+};
 
-export const WithIcon = Template.bind({});
+export const WithChain = Template.bind({});
 // More on args: https://storybook.js.org/docs/react/writing-stories/args
-WithIcon.args = {
-  isShowSubIcon: true,
+WithChain.args = {
+  rightComponent: (
+    <div style={{ paddingRight: 8 }}>
+      <Logo size={24} network="polkadot" />
+    </div>
+  ),
   subIcon: <Icon type="phosphor" phosphorIcon={QrCode} iconColor="#FFF" size="xs" weight="bold" />,
 };
